% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAPI_code.R
\name{add_endpoint_}
\alias{add_endpoint_}
\title{add_endpoint_}
\usage{
add_endpoint_(x = add_ref_(), y = "hsub")
}
\arguments{
\item{x}{dataframe}

\item{y}{string}
}
\description{
list of relevant static parsed urls, accessible by alias
}
