% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAPI_code.R
\name{get_json_}
\alias{get_json_}
\title{get_json_}
\usage{
get_json_(
  x = c("gal2:19", "rom3:17", "ROM2:13!SG21", "MAT7:21!SG21", "JAM1:22!SG21",
    "LUK6:46!SG21")
)
}
\arguments{
\item{x}{string}
}
\description{
list of relevant static parsed urls, accessible by alias
}
