% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAPI_code.R
\name{render_refs_}
\alias{render_refs_}
\title{render_refs_}
\usage{
render_refs_(
  x = c("ROM2:13!SG21", "MAT7:21!SG21", "JAM1:22!SG21", "LUK6:46!SG21", "HEB4:2!SG21",
    "1SA3:10!SG21", "1SA3:1!SG21", "PSA33:9!SG21", "ROM1:19-20!SG21")
)
}
\arguments{
\item{x}{dataframe}
}
\description{
list of relevant static parsed urls, accessible by alias
}
